move fast color lookup to its own class
