+++ Assets/Scripts/Utility/FastColorPalette.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 98d817ecf6f2243ba83f987c50353fd9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
